requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
flask==3.1.0
flask-cors==5.0.1
werkzeug==3.1.3
kafka-python==2.0.2

